requirements.txt
flask
Frozen-Flask
gitpython
markdown